fix(e2e): move types to prevent collisions with upstream equivalent
