+++ workspaces/backstage/plugins/techdocs/overlay/src/utils.tsx
@@ -1,4 +1,16 @@
-import { DynamicConfig, TechdocsAddon } from './types';
+type TechdocsAddon = {
+  scope: string;
+  module: string;
+  importName: string;
+  Component: React.ComponentType<React.PropsWithChildren>;
+  config: {
+    props?: Record<string, any>;
+  };
+};
+
+export type DynamicConfig = {
+  techdocsAddons: TechdocsAddon[];
+};
 
 function getTechdocsAddonData(dynamicConfig: DynamicConfig): TechdocsAddon[] {
   return dynamicConfig?.techdocsAddons ?? [];
